+++ src/main/java/de/jeff_media/angelchest/config/Config.java
@@ -109,4 +109,5 @@ public final class Config {
     public static final String ALLOW_FETCH_ACROSS_WORLDS = "allow-fetch-across-worlds";
     public static final String MAX_TP_DISTANCE = "max-tp-distance";
     public static final String MAX_FETCH_DISTANCE = "max-fetch-distance";
+    public static final String EXEMPT_ELITEMOBS_SOULBOUND_ITEMS_FROM_GENERIC_SOULBOUND_DETECTION = "exempt-elitemobs-soulbound-items-from-generic-soulbound-detection";
 }